tidy file usecase
